requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-apscheduler==0.6.0
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
mysqlclient==2.1.0
PyJWT==2.3.0
pyotp==2.6.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
